# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reserva de Laboratórios e Salas",
    "version": "1.0.0",
    "description": "Importe no Insomnia. Cadastre usuário, confira login e configure Basic Auth. Use IDs retornados nas consultas. Datas DD/MM/AAAA e horários HH:MM:SS. Testes ainda pendentes."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
    "/api/usuarios": {
      "get": {
        "summary": "Consultar usuarios",
        "tags": [
          "usuarios"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cpf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataAniversario",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
              "example": "30/09/2026",
              "description": "DD/MM/AAAA; use uma data futura na reserva"
            }
          }
        ]
      },
      "post": {
        "summary": "Cadastrar usuarios",
        "tags": [
          "usuarios"
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioView"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URI do registro criado",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioInput"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/usuarios/{codigo}": {
      "get": {
        "summary": "Buscar por código",
        "tags": [
          "usuarios"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "put": {
        "summary": "Atualizar cadastro completo",
        "tags": [
          "usuarios"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "delete": {
        "summary": "Excluir registro (reserva ativa respeita prazo de cancelamento)",
        "tags": [
          "usuarios"
        ],
        "responses": {
          "204": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "UsuarioInput": {
        "type": "object",
        "properties": {
          "cpf": {
            "type": "string",
            "minLength": 11,
            "maxLength": 11,
            "example": "52998224725"
          },
          "nomeCompleto": {
            "type": "string",
            "minLength": 10,
            "maxLength": 80,
            "example": "Professor de Teste"
          },
          "dataAniversario": {
            "type": "string",
            "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
            "example": "15/03/1990",
            "description": "DD/MM/AAAA; use uma data futura na reserva"
          },
          "celular": {
            "type": "string",
            "maxLength": 20,
            "example": "11999999999"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 15,
            "maxLength": 80,
            "example": "professor.teste@example.com"
          },
          "login": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._-]{3,80}$",
            "example": "professor.teste"
          },
          "senha": {
            "type": "string",
            "minLength": 6,
            "maxLength": 60,
            "format": "password",
            "example": "Teste123!"
          }
        },
        "required": [
          "cpf",
          "nomeCompleto",
          "dataAniversario",
          "celular",
          "email",
          "login",
          "senha"
        ],
        "additionalProperties": false
      },
      "UsuarioView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "cpf": {
            "type": "string",
            "minLength": 11,
            "maxLength": 11,
            "example": "52998224725"
          },
          "nomeCompleto": {
            "type": "string",
            "minLength": 10,
            "maxLength": 80,
            "example": "Professor de Teste"
          },
          "dataAniversario": {
            "type": "string",
            "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
            "example": "15/03/1990",
            "description": "DD/MM/AAAA; use uma data futura na reserva"
          },
          "celular": {
            "type": "string",
            "maxLength": 20,
            "example": "11999999999"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 15,
            "maxLength": 80,
            "example": "professor.teste@example.com"
          },
          "login": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._-]{3,80}$",
            "example": "professor.teste"
          }
        },
        "required": [
          "codigo",
          "cpf",
          "nomeCompleto",
          "dataAniversario",
          "celular",
          "email",
          "login"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "mensagem": {
            "type": "string"
          },
          "caminho": {
            "type": "string"
          },
          "erros": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "campo": {
                  "type": "string"
                },
                "mensagem": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensagem"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "mensagem",
          "caminho",
          "erros"
        ],
        "additionalProperties": false
      }
    }
  }
}